+++ application-bemyplan/src/main/java/com/deploy/bemyplan/service/scrap/ScrapService.java
@@ -22,20 +22,13 @@ public class ScrapService {
 
     @Transactional
     public void deleteScrap(Long userId, Long planId){
-        Scrap scrap = scrapRepository.findActiveByUserIdAndPlanId(planId, userId);
-        if (scrap == null) {
-            throw new NotFoundException(String.format("해당 여행일정을 찜한 상태 (%s - %s) 가 아닙니다.", userId, planId), NOT_FOUND_SCRAP_EXCEPTION);
-        }
-        scrap.updateToInActive();
-        scrapRepository.save(scrap);
+        scrapRepository.deleteByUserIdAndPlanId(userId, planId);
     }
 
     @Transactional(readOnly = true)
     public void checkScrapStatus(Long planId, Long userId) {
-        Scrap scrap = scrapRepository.findActiveByUserIdAndPlanId(planId, userId);
-        if (scrap == null) {
-            throw new NotFoundException(String.format("해당 여행일정을 찜한 상태 (%s - %s) 가 아닙니다.", userId, planId), NOT_FOUND_SCRAP_EXCEPTION);
-        }
+        scrapRepository.findByUserIdAndPlanId(planId, userId)
+                .orElseThrow(() -> new NotFoundException(String.format("해당 여행일정을 찜한 상태 (%s - %s) 가 아닙니다.", userId, planId), NOT_FOUND_SCRAP_EXCEPTION));
     }
 }
