+++ module-domain-bemyplan/src/main/java/com/deploy/bemyplan/domain/scrap/repository/ScrapRepositoryCustom.java
@@ -9,6 +9,5 @@ public interface ScrapRepositoryCustom {
 
     List<Scrap> findAllByIds(List<Long> scrapIds);
     List<Scrap> findByUserIdAndPlanIds(List<Long> planIds, @Nullable Long userId);
-    Scrap findByUserIdAndPlanId(Long planId, @Nullable Long userId);
     Scrap findActiveByUserIdAndPlanId(Long planId, @Nullable Long userId);
 }
